api/requirements.txt
asgiref==3.7.2
Django==4.2.7
django-cors-headers==4.3.0
django-filter==23.3
djangorestframework==3.14.0
drf-yasg==1.21.7
factory-boy==3.3.0
Faker==19.13.0
inflection==0.5.1
packaging==23.2
Pillow==10.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
uritemplate==4.1.1
